Fix crash safety for qsbr_per_thread::qsbr_resume
Move allocation to be the first thing in qsbr_resume, so that throwing an
exception there will not result in any other changes.

Start QSBR OOM tests in the new file test/test_qsbr_oom.cpp. For that, extract
GTest test base class into new helper files qsbr_test_utils.hpp &
qsbr_test_utils.cpp.
